feat: splash mvi 패턴 적용
