+++ lib/core/routing/router.dart
@@ -9,7 +9,7 @@ import 'package:recipe_app/presentation/login/login_screen.dart';
 import 'package:recipe_app/presentation/register/register_screen.dart';
 import 'package:recipe_app/presentation/saved_recipes/saved_recipes_screen.dart';
 import 'package:recipe_app/presentation/search_recipes/search_recipes_screen.dart';
-import 'package:recipe_app/presentation/splash/splash_screen.dart';
+import 'package:recipe_app/presentation/splash/splash_screen_root.dart';
 
 final _rootNavigatorKey = GlobalKey<NavigatorState>();
 final _shellNavigatorKey = GlobalKey<NavigatorState>();
@@ -20,7 +20,7 @@ final router = GoRouter(
   routes: [
     GoRoute(
       path: Routes.splash,
-      builder: (context, state) => const SplashScreen(),
+      builder: (context, state) => const SplashScreenRoot(),
     ),
     GoRoute(
       path: Routes.login,
